+++ weblate/trans/permissions.py
@@ -368,3 +368,11 @@ def can_add_translation(user, project):
     Checks whether user can view reports on given project.
     """
     return check_permission(user, project, 'trans.add_translation')
+
+
+@cache_permission
+def can_remove_translation(user, project):
+    """
+    Checks whether user can view reports on given project.
+    """
+    return check_permission(user, project, 'trans.remove_translation')